Add files via upload
A comprehensive resource for biologists and computational scientists to delve into the world of bioimage analysis. This repository offers a practical guide to classical computer vision techniques, with a particular focus on optical flow methods. Explore the power of Lucas-Kanade optical flow to track the movement of biological objects, such as cells, organelles, or fluorescently labeled proteins, within microscopic image sequences.